# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nimiq Simple Faucet
  version: 0.0.1
  description: Public, admin, and MCP surfaces for the Nimiq Simple Faucet
    (testnet). See /SECURITY.md for the threat model.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Security
    url: https://github.com/onmax/Nimiq-Simple-Faucet/blob/main/SECURITY.md
servers:
  - url: /
    description: This instance
...
components:
  securitySchemes:
    integratorHmac:
      type: apiKey
      in: header
      name: X-Faucet-Api-Key
      description: "Integrator HMAC auth: include `X-Faucet-Api-Key`,
        `X-Faucet-Timestamp`, `X-Faucet-Nonce`, and `X-Faucet-Signature`
        (base64-HMAC-SHA256 over
        `method\\npath\\ntimestamp\\nnonce\\nbodyHash`)."
    adminSession:
      type: apiKey
      in: cookie
      name: faucet_session
      description: Session cookie issued by POST /admin/auth/login. Mutating calls
        also require the `X-Faucet-Csrf` double-submit header matching the
        `faucet_csrf` cookie.
    adminMcpToken:
      type: apiKey
      in: header
      name: X-Faucet-Admin-Token
      description: Required to invoke admin-scoped MCP tools over POST /mcp.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        issues:
          type: array
          items: {}
      required:
        - error
    LoginRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 512
        totp:
          type: string
          minLength: 1
          maxLength: 16
      required:
        - password
    LoginResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - true
        expiresAt:
          type: string
        totpProvisioningUri:
          type: string
        totpSecret:
          type: string
      required:
        - ok
        - expiresAt
...
    BlocklistEntry:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          enum:
            - ip
            - address
            - uid
            - asn
            - country
        value:
          type: string
        reason:
          type:
            - string
            - "null"
        createdAt:
          anyOf:
            - type: string
            - type: number
        expiresAt:
          anyOf:
            - type: string
            - type: number
            - type: "null"
      required:
        - id
        - kind
        - value
        - reason
        - createdAt
        - expiresAt
    BlocklistCreateRequest:
      type: object
      properties:
        kind:
          type: string
          enum:
            - ip
            - address
            - uid
            - asn
            - country
        value:
          type: string
          minLength: 1
          maxLength: 128
        reason:
          type: string
          maxLength: 256
        expiresAt:
          anyOf:
            - type: string
            - type: number
      required:
        - kind
        - value
...
    AdminAccount:
      type: object
      properties:
        address:
          type: string
        balance:
          type: string
        recentPayouts:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              address:
                type: string
              amountLuna:
                type: string
              txId:
                type:
                  - string
                  - "null"
              createdAt:
                anyOf:
                  - type: string
                  - type: number
              status:
                type: string
            required:
              - id
              - address
              - amountLuna
              - txId
              - createdAt
              - status
      required:
        - address
        - balance
        - recentPayouts
    AdminSendRequest:
      type: object
      properties:
        to:
          type: string
          minLength: 1
        amountLuna:
          anyOf:
            - type: string
              pattern: ^\d+$
            - type: integer
              minimum: 1
        memo:
          type: string
          maxLength: 256
      required:
        - to
        - amountLuna
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - true
      required:
        - ok
...
paths:
...
  /admin/auth/login:
    post:
      tags:
        - Admin
      summary: Password (+ TOTP) login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/logout:
    post:
      tags:
        - Admin
      summary: Revoke the current admin session
      security:
        - adminSession: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /admin/blocklist:
    get:
      tags:
        - Admin
      summary: List blocklist entries
      security:
        - adminSession: []
      responses:
        "200":
          description: Entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/BlocklistEntry"
                required:
                  - total
                  - items
    post:
      tags:
        - Admin
      summary: Add a blocklist entry
      security:
        - adminSession: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlocklistCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                required:
                  - id
...
  /admin/account:
    get:
      tags:
        - Admin
      summary: Faucet wallet account info
      security:
        - adminSession: []
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccount"
  /admin/account/send:
    post:
      tags:
        - Admin
      summary: Manual send (requires TOTP step-up)
      security:
        - adminSession: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSendRequest"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  txId:
                    type: string
                required:
                  - txId
  /admin/account/rotate-key:
    post:
      tags:
        - Admin
      summary: Rotate encrypted at-rest key material (requires TOTP step-up)
      security:
        - adminSession: []
      responses:
        "200":
          description: Rotation timestamp
          content:
            application/json:
              schema:
                type: object
                properties:
                  rotatedAt:
                    type: string
                required:
                  - rotatedAt
...
  /admin/blocklist/{id}:
    delete:
      tags:
        - Admin
      summary: Remove a blocklist entry
      security:
        - adminSession: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/auth/totp/enroll:
    post:
      tags:
        - Admin
      summary: Pre-enrolment TOTP provisioning (first-time admin setup)
      description: Returns a TOTP secret and provisioning URI. Only usable before any
        admin user exists.
      responses:
        "200":
          description: TOTP provisioning
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  provisioningUri:
                    type: string
                required:
                  - secret
                  - provisioningUri
        "409":
          description: Already enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/auth/reset:
    post:
      tags:
        - Admin
      summary: "Dev-only: wipe all faucet state (password-protected)"
      description: Requires FAUCET_DEV=1 and FAUCET_ADMIN_PASSWORD in the request
        body. Returns 404 in production.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required:
                - password
      responses:
        "200":
          description: State wiped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Wrong or missing password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not available (production mode)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...